+++ merlin/main.py
@@ -270,7 +270,7 @@ def config_merlin(args):
     if output_dir is None:
         USER_HOME = os.path.expanduser("~")
         output_dir = os.path.join(USER_HOME, ".merlin")
-    _ = router.create_config(args.task_server, output_dir)
+    _ = router.create_config(args.task_server, output_dir, args.broker)
 
 
 def process_example(args):
@@ -559,6 +559,13 @@ def setup_argparse():
         default=None,
         help="Optional directory to place the default config file.\
                             Default: ~/.merlin",
+    )
+    mconfig.add_argument(
+        "--broker",
+        type=str,
+        default=None,
+        help="Optional broker type, backend will be redis\
+                            Default: rabbitmq",
     )
 
     # merlin example